backend/requirements.txt
# NerdsIQ Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# AI/ML - RAG Pipeline
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.19
openai>=1.12.0
tiktoken>=0.6.0

# Vector Database
qdrant-client>=1.7.0

# Google Drive Integration
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.0
